moved test to new LatestIT
